- Do not close the status dialog after the session is over. We disconnect the process and leave the user with information about what operations were performed on his computer

+++ host/host_session_console.cc
@@ -137,7 +137,11 @@ void HostSessionConsole::OnSessionDetached()
 
     process_watcher_.StopWatching();
 
-    process_.Terminate(0, false);
+    if (session_type_ != proto::SessionType::SESSION_TYPE_FILE_TRANSFER)
+    {
+        process_.Terminate(0, false);
+    }
+
     process_.Close();
 
     ipc_channel_.reset();
